+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to KoviD will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## [Unreleased]
+### Fixed
+- Fix `this_mod->state` being overridden in `do_init_module` by kernel
+
+### Added
+- Stealth: add `mod_tree` hiding
+
 ## v4.0.3
 ### Fixed
 - Fix build regression regarding kernel 6.14